% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{as.matrix.profile}
\alias{as.matrix.profile}
\title{Convert a tsmp object in another if possible}
\usage{
\method{as.matrix}{profile}(.mp)
}
\description{
Convert a tsmp object in another if possible
}
\keyword{internal}
