% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{as.multimatrix.profile}
\alias{as.multimatrix.profile}
\title{Convert a tsmp object in another if possible}
\usage{
as.multimatrix.profile(.mp)
}
\description{
Convert a tsmp object in another if possible
}
\keyword{internal}
