% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{av.apply}
\alias{av.apply}
\title{Corrects the matrix profile using an annotation vector}
\usage{
av.apply(matrix.profile, annotation.vector)
}
\arguments{
\item{matrix.profile}{The matrix profile.}

\item{annotation.vector}{The annotation vector.}
}
\value{
Returns the corrected matrix profile
}
\description{
Corrects the matrix profile using an annotation vector
}
\examples{
av <- av.complexity(data, window)
mpc <- av.apply(mp, av)

}
\references{
\itemize{
\item Dau HA, Keogh E. Matrix Profile V: A Generic Technique to Incorporate Domain
Knowledge into Motif Discovery. In: Proceedings of the 23rd ACM SIGKDD International Conference
on Knowledge Discovery and Data Mining - KDD ’17. New York, New York, USA: ACM Press; 2017. p.
125–34.
}
}
