% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{av.complexity}
\alias{av.complexity}
\title{Computes the annotation vector that favors complexity}
\usage{
av.complexity(data, window.size, dilution.factor = 0)
}
\arguments{
\item{data}{a \code{vector} or a column \code{matrix} of \code{numeric}.}

\item{window.size}{an \code{int}. Size of the sliding window.}

\item{dilution.factor}{a \code{numeric}. (Default is \code{0}). Larger numbers means more dilution.}
}
\value{
Returns the annotation vector for matrix profile correction.
}
\description{
Computes the annotation vector that favors complexity
}
\examples{
data <- test_data$train$data[1:1000]
window <- 50
av <- av.complexity(data, window)

}
\references{
\itemize{
\item Dau HA, Keogh E. Matrix Profile V: A Generic Technique to Incorporate Domain
Knowledge into Motif Discovery. In: Proceedings of the 23rd ACM SIGKDD International Conference
on Knowledge Discovery and Data Mining - KDD ’17. New York, New York, USA: ACM Press; 2017. p.
125–34.
}
}
