% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{av.hardlimit.artifact}
\alias{av.hardlimit.artifact}
\title{Computes the annotation vector that suppresses hard-limited artifacts}
\usage{
av.hardlimit.artifact(data, window.size)
}
\arguments{
\item{data}{a \code{vector} or a column \code{matrix} of \code{numeric}.}

\item{window.size}{an \code{int}. Size of the sliding window.}
}
\value{
Returns the annotation vector for matrix profile correction.
}
\description{
Computes the annotation vector that suppresses hard-limited artifacts
}
\examples{
data <- test_data$train$data[1:1000]
window <- 50
av <- av.hardlimit.artifact(data, window)

}
\references{
\itemize{
\item Dau HA, Keogh E. Matrix Profile V: A Generic Technique to Incorporate Domain
Knowledge into Motif Discovery. In: Proceedings of the 23rd ACM SIGKDD International Conference
on Knowledge Discovery and Data Mining - KDD ’17. New York, New York, USA: ACM Press; 2017. p.
125–34.
}
}
