% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{av_motion_artifact}
\alias{av_motion_artifact}
\title{Computes the annotation vector that suppresses motion artifacts}
\usage{
av_motion_artifact(.mp, data, apply = FALSE)
}
\arguments{
\item{data}{a \code{vector} or a column \code{matrix} of \code{numeric}.}
}
\value{
Returns the annotation vector for matrix profile correction.
}
\description{
Computes the annotation vector that suppresses motion artifacts
}
\examples{
data <- mp_test_data$train$data[1:1000]
window <- 50
av <- av_motion_artifact(data, window)

}
\references{
\itemize{
\item Dau HA, Keogh E. Matrix Profile V: A Generic Technique to Incorporate Domain
Knowledge into Motif Discovery. In: Proceedings of the 23rd ACM SIGKDD International Conference
on Knowledge Discovery and Data Mining - KDD ’17. New York, New York, USA: ACM Press; 2017. p.
125–34.
}
}
