% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds.R
\name{bitsave.sub.picking}
\alias{bitsave.sub.picking}
\title{Title}
\usage{
bitsave.sub.picking(data, matrix.profile, profile.index, window.size, bits,
  n.cand, exclusion.zone = 1/2, verbose = 2)
}
\arguments{
\item{exclusion.zone}{}
}
\description{
Title
}
\examples{
distMat <- as.matrix(dist(carfull$data))
matrixProfile <- matrix(0, nrow(carfull$data), 1)
profileIndex <- matrix(0, nrow(carfull$data), 1)

for (i in 1:nrow(carfull$data)) {
  distMat[i, i] <- Inf;
  profileIndex[i] <- which.min(distMat[i,])
  matrixProfile[i] <- distMat[i,profileIndex[i]]
}
}
\references{
\enumerate{
\item Yeh CCM, Van Herle H, Keogh E. Matrix profile III: The matrix profile allows visualization of salient subsequences in massive time series. Proc - IEEE Int Conf Data Mining, ICDM. 2017;579–88.
}

\enumerate{
\item Hu B, Rakthanmanon T, Hao Y, Evans S, Lonardi S, Keogh E. Discovering the Intrinsic Cardinality and Dimensionality of Time Series Using MDL. In: 2011 IEEE 11th International Conference on Data Mining. IEEE; 2011. p. 1086–91.
}

Website: \url{https://sites.google.com/site/salientsubs/}
}
