% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{complexity}
\alias{complexity}
\title{Computes the complexity index of the data}
\usage{
complexity(data)
}
\arguments{
\item{data}{a \code{vector} or a column \code{matrix} of \code{numeric}.}
}
\value{
Returns the complexity index of the data provided (normally a subset)
}
\description{
Computes the complexity index of the data
}
\keyword{internal}
