% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdts_train.R
\name{compute.f.meas}
\alias{compute.f.meas}
\title{Title}
\usage{
compute.f.meas(label, pos.st, pos.ed, dist.pro, thold, window.size, beta)
}
\arguments{
\item{beta}{}
}
\description{
Title
}
\keyword{internal}
