% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{diff2}
\alias{diff2}
\title{Distance between two matrices}
\usage{
diff2(x, y)
}
\arguments{
\item{x}{a \code{matrix}.}

\item{y}{a \code{matrix}.}
}
\value{
Returns a \code{matrix} of size m x n if x is of size m x k and y is of size n x k.
}
\description{
Computes the Euclidean distance between rows of two matrices.
}
\keyword{internal}
