% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salient.R
\name{discrete.norm.pre}
\alias{discrete.norm.pre}
\title{Precompute the max and min value for the discrete normalization}
\usage{
discrete.norm.pre(data, window.size = 1)
}
\arguments{
\item{data}{input time series}

\item{window.size}{sliding window size}
}
\value{
Returns a list with the max and min value
}
\description{
Precompute the max and min value for the discrete normalization
}
\keyword{internal}
