% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salient.R
\name{discrete.norm}
\alias{discrete.norm}
\title{Discrete normalization}
\usage{
discrete.norm(data, n.bits, max, min)
}
\arguments{
\item{data}{Input time series.}

\item{n.bits}{Number of bits for MDL discretization.}

\item{max}{Precomputed max from \code{discrete.norm.pre}.}

\item{min}{Precomputed min from \code{discrete.norm.pre}.}
}
\value{
Returns the data after discrete normalization.
}
\description{
Discrete normalization
}
\keyword{internal}
