% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{fast_movavg}
\alias{fast_movavg}
\title{Fast implementation of moving average}
\usage{
fast_movavg(data, window_size)
}
\arguments{
\item{data}{a \code{vector} or a column \code{matrix} of \code{numeric}.}

\item{window_size}{moving sd window size}
}
\value{
Returns a \code{vector} with the moving average
}
\description{
This function does not handle NA values
}
\examples{
data_avg <- fast_movavg(mp_toy_data$data[, 1], mp_toy_data$sub_len)
}
