% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{fast_movsd}
\alias{fast_movsd}
\title{Fast implementation of moving standard deviation using filter}
\usage{
fast_movsd(data, window_size)
}
\arguments{
\item{data}{a \code{vector} or a column \code{matrix} of \code{numeric}.}

\item{window_size}{moving sd window size}
}
\value{
Returns a \code{vector} with the moving standard deviation
}
\description{
Fast implementation of moving standard deviation using filter
}
\examples{
data_sd <- fast_movsd(mp_toy_data$data[, 1], mp_toy_data$sub_len)
}
