% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_chains.R
\name{find.chains}
\alias{find.chains}
\title{Find Time Series Chains}
\usage{
find.chains(matrices)
}
\arguments{
\item{matrices}{a result from STAMP or STOMP algorithms}
}
\value{
Returns \code{chains}, a \code{list} of chains founded with more than 2 patterns and \code{best}
with the best one.
}
\description{
Find Time Series Chains
}
\examples{
w <- 50
data <- gait_data
mp <- stamp(data, window.size = w, exclusion.zone = 1/4, verbose = 0)
find.chains(mp)

}
\references{
\enumerate{
\item Zhu Y, Imamura M, Nikovski D, Keogh E. Introducing time series chains: a new
primitive for time series data mining. Knowl Inf Syst. 2018 Jun 2;1–27.
}

Website: \url{https://sites.google.com/site/timeserieschain/}
}
