% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_motifs.R
\name{find.motifs}
\alias{find.motifs}
\title{Find MOTIFs and Plot them}
\usage{
find.motifs(matrix.profile, profile.index, window.size, data,
  exclusion.zone = 1/2, radius = 3, n.motifs = 3, plot = TRUE,
  cols = 3)
}
\arguments{
\item{cols}{}
}
\description{
Find MOTIFs and Plot them
}
