% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_motifs.R
\name{find_motif.MatrixProfile}
\alias{find_motif.MatrixProfile}
\title{Title}
\usage{
\method{find_motif}{MatrixProfile}(.mp, data, n_motifs = 3, radius = 3,
  exclusion_zone = NULL)
}
\description{
Title
}
