% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_motifs.R
\name{find_motif.MultiMatrixProfile}
\alias{find_motif.MultiMatrixProfile}
\title{Title}
\usage{
\method{find_motif}{MultiMatrixProfile}(.mp, data, n_motifs = 3,
  mode = c("guided", "unguided"), n_bit = 4, exclusion_zone = NULL,
  n_dim = NULL)
}
\description{
Title
}
