% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_motifs.R
\name{find_motifs}
\alias{find_motifs}
\title{Find MOTIFs}
\usage{
find_motifs(.mp, data, n_motifs = 3, radius = 3, mode = c("guided",
  "unguided"), n_bit = 4, exclusion_zone = NULL, n_dim = NULL)
}
\arguments{
\item{n_dim}{}
}
\description{
Find MOTIFs
}
