% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-snippet.R
\name{find_snippet}
\alias{find_snippet}
\title{Time Series Snippets: A New Primitive for Time Series Data Mining}
\usage{
find_snippet(data, s_size, n_snippets = 2, window_size = s_size/2)
}
\arguments{
\item{data}{a \code{matrix} or a \code{vector}.}

\item{s_size}{an int. Size of snippet.}

\item{n_snippets}{an \code{int}. Number of snippets to find. (Default is \code{2}).}

\item{window_size}{an \code{int}. The size of the sliding window used to compare the data. Must be smaller
than \code{s_size}. (Default is \code{s_size / 2}).}
}
\value{
Returns the snippet : a list of n_snippets snippets
fraction : fraction of each snippet
snippetidx : the location of each
snippet within time sereis
}
\description{
MPdist is a recently introduced distance measure which considers two time series to be similar
}
\details{
MPdist
}
\examples{

}
\references{
\itemize{
\item Imani S, Madrid F, Ding W, Crouter S, Keogh E. Matrix Profile XIII:
Time Series Snippets: A New Primitive for Time Series Data Mining. In: 2018 IEEE International
Conference on Data Mining (ICDM). 2018.
}

\itemize{
\item Gharghabi S, Imani S, Bagnall A, Darvishzadeh A, Keogh E. Matrix Profile XII:
MPdist: A Novel Time Series Distance Measure to Allow Data Mining in More Challenging Scenarios.
In: 2018 IEEE International Conference on Data Mining (ICDM). 2018.
}

Website: \url{https://sites.google.com/site/snippetfinder/}
}
