% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluss.R
\name{fluss.extract}
\alias{fluss.extract}
\title{FLUSS - Extract Segments}
\usage{
fluss.extract(arc.counts, num.segments, window.size, exclusion.zone = 5)
}
\arguments{
\item{arc.counts}{a \code{matrix} with the corrected arc counts from \code{\link[=fluss.cac]{fluss.cac()}}.}

\item{num.segments}{an \code{int}. Number of segments to extract. Based on domain knowledge.}

\item{window.size}{an \code{int}. Size of the sliding window.}

\item{exclusion.zone}{a \code{numeric}. Size of the exclusion zone, based on window size (default is \code{5}).}
}
\value{
Returns an \code{int} or a \code{vector} of \code{int} with the location of predicted semantic changes.
The number of locations is not greater than \code{num.segments}.
}
\description{
Extract candidate points of semantic changes.
}
\examples{
# This is a fast toy example and results are useless. For a complete result, run the code inside
#'Not run' section below.
data <- fluss_data$tilt.abp$data[1:1000]
w <- 210
mp <- mstomp(data, w, verbose = 0)
cac <- fluss.cac(mp$pi, w)
segments <- fluss.extract(cac, 1, w)
\dontrun{
data <- fluss_data$walkjogrun$data
w <- fluss_data$walkjogrun$window # 80
nseg <- length(fluss_data$walkjogrun$gtruth) # 2
mp <- mstomp(data, w)
cac <- fluss.cac(mp$pi, w)
segments <- fluss.extract(cac, nseg, w)
}
}
\references{
\enumerate{
\item Gharghabi S, Ding Y, Yeh C-CM, Kamgar K, Ulanova L, Keogh E. Matrix Profile VIII: Domain Agnostic Online Semantic Segmentation at Superhuman Performance Levels. In: 2017 IEEE International Conference on Data Mining (ICDM). IEEE; 2017. p. 117–26.
}

Website: \url{https://sites.google.com/site/onlinesemanticsegmentation/}

Website: \url{http://www.cs.ucr.edu/~eamonn/MatrixProfile.html}
}
\seealso{
Other fluss: \code{\link{fluss.cac}},
  \code{\link{fluss.score}}, \code{\link{fluss}}
}
\concept{fluss}
