% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gait_data.R
\docType{data}
\name{gait_data}
\alias{gait_data}
\title{Original data used in the Time Series Chain demo}
\format{A \code{matrix} with 904 rows and 1 column with the Y data from an accelerometer}
\source{
\url{https://sites.google.com/site/timeserieschain/}
}
\usage{
gait_data
}
\description{
Original data used in the Time Series Chain demo
}
\references{
\enumerate{
\item Zhu Y, Imamura M, Nikovski D, Keogh E. Introducing time series chains: a new primitive for time series data mining. Knowl Inf Syst. 2018 Jun 2;1–27.
}
}
\keyword{datasets}
