% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds.R
\name{get.mds}
\alias{get.mds}
\title{Future function to see MDS}
\usage{
get.mds(data, sub.picking, window.size)
}
\arguments{
\item{data}{original data}

\item{sub.picking}{picked subsequences}

\item{window.size}{window size}
}
\value{
Returns X,Y values for plotting
}
\description{
Future function to see MDS
}
