% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salient.R
\name{get.sorted.idx}
\alias{get.sorted.idx}
\title{Retrieve the index of a number of candidates from the lowest points of a MP}
\usage{
get.sorted.idx(matrix.profile, n.cand, exclusion.zone = 0)
}
\arguments{
\item{matrix.profile}{the matrix profile}

\item{n.cand}{number of candidates to extract}

\item{exclusion.zone}{exclusion zone for extracting candidates (in absolute values)}
}
\value{
Returns the indexes of candidates
}
\description{
Retrieve the index of a number of candidates from the lowest points of a MP
}
\keyword{internal}
