% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_data}
\alias{get_data}
\title{Get the data included in a TSMP object, if any.}
\usage{
get_data(.mp)
}
\arguments{
\item{.mp}{a TSMP object.}
}
\value{
Returns the data as \code{matrix}. If there is more than one series, returns a \code{list}.
}
\description{
Get the data included in a TSMP object, if any.
}
\examples{
mp <- tsmp(mp_toy_data$data[1:200, 1], window_size = 30, verbose = 0)
get_data(mp)
}
