% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdts_train.R
\name{golden.section.2}
\alias{golden.section.2}
\title{Computes the golden section for combined candidates}
\usage{
golden.section.2(dist.pro, thold, label, pos.st, pos.ed, beta, window.size,
  fit.idx)
}
\arguments{
\item{dist.pro}{the candidate distance profile}

\item{thold}{a number with the threshold used to calculate the F-Score}

\item{label}{a vector with the data bool annotation}

\item{pos.st}{a vector with the starting points of label}

\item{pos.ed}{a vector with the ending points of label}

\item{beta}{a number that balance the F-Score. Beta > 1 towards recall, < towards precision}

\item{window.size}{an integer with the sliding window size}

\item{fit.idx}{an integer with the index of the current threshold}
}
\value{
Returns the best threashold and its F-Score
}
\description{
Computes the golden section for combined candidates
}
\keyword{internal}
