% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_guide_search.R
\name{guide_search}
\alias{guide_search}
\title{Guided MOTIF Discovery for Multidimensional Matrix Profile}
\usage{
guide_search(data, window_size, matrix_profile, profile_index, n_dim)
}
\arguments{
\item{data}{a \code{matrix} of \code{numeric}, where each column is a time series. Accepts \code{vector} (see
details), \code{list} and \code{data.frame} too.}

\item{window_size}{an \code{int} with the size of the sliding window.}

\item{matrix_profile}{multidimensional matrix profile (from \code{\link[=mstomp]{mstomp()}} or \code{\link[=mstomp_par]{mstomp_par()}}).}

\item{profile_index}{multidimensional profile index (from \code{\link[=mstomp]{mstomp()}} or \code{\link[=mstomp_par]{mstomp_par()}}).}

\item{n_dim}{an \code{int}. The dimensionality of the MOTIF to find.}
}
\value{
Returns the \code{motif_idx} with the index of MOTIFs founded and \code{motif_dim} with the spanned
dimensions of respective MOTIF.
}
\description{
Guided MOTIF Discovery for Multidimensional Matrix Profile
}
\details{
Although this functions handles Multivariate Time Series, it can also be used to handle
Univariate Time Series.
}
\examples{
# This is a fast toy example and results are useless. For a complete result, run the code inside
#'Not run' section below.
w <- mp_toy_data$sub_len
mp <- mstomp(mp_toy_data$data[1:200,], w, verbose = 0)
motifs <- guide_search(mp_toy_data$data[1:200,], w, mp$mp, mp$pi, 2)
\dontrun{
w <- mp_toy_data$sub_len
mp <- mstomp_par(mp_toy_data$data, w, verbose = 0)
motifs <- guide_search(mp_toy_data$data, w, mp$mp, mp$pi, 2)
}
}
\references{
\itemize{
\item Yeh CM, Kavantzas N, Keogh E. Matrix Profile VI : Meaningful Multidimensional Motif
Discovery.
}

\itemize{
\item Zhu Y, Imamura M, Nikovski D, Keogh E. Matrix Profile VII: Time Series Chains: A
New Primitive for Time Series Data Mining. Knowl Inf Syst. 2018 Jun 2;1–27.
}

Website: \url{https://sites.google.com/view/mstamp/}

Website: \url{http://www.cs.ucr.edu/~eamonn/MatrixProfile.html}
}
\seealso{
\code{\link[=mstomp]{mstomp()}}, \code{\link[=mstomp_par]{mstomp_par()}}, \code{\link[=unconstrain_search]{unconstrain_search()}}
}
