% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_motif_usearch.R, R/m_unconstrain_search.R
\name{m_motif_usearch}
\alias{m_motif_usearch}
\title{MDL Based MOTIF Discovery for Multidimensional Matrix Profile}
\usage{
m_motif_usearch(data, window_size, matrix_profile, profile_index,
  n_bit = 4, k = Inf)

m_motif_usearch(data, window_size, matrix_profile, profile_index,
  n_bit = 4, k = Inf)
}
\arguments{
\item{data}{a \code{matrix} of \code{numeric}, where each column is a time series. Accepts \code{vector} (see
details), \code{list} and \code{data.frame} too.}

\item{window_size}{an \code{int} with the size of the sliding window.}

\item{matrix_profile}{multidimensional matrix profile (from \code{\link[=mstomp]{mstomp()}} or \code{\link[=mstomp_par]{mstomp_par()}}).}

\item{profile_index}{multidimensional profile index (from \code{\link[=mstomp]{mstomp()}} or \code{\link[=mstomp_par]{mstomp_par()}}).}

\item{n_bit}{an \code{int}. Number of bits for MDL discretization. (Default is \code{4}).}

\item{k}{an \code{int}. The number of MOTIFs to retrieve. \code{Inf} means all possible MOTIFs. (Default is
\code{Inf}).}

\item{data}{a \code{matrix} of \code{numeric}, where each column is a time series. Accepts \code{vector} (see
details), \code{list} and \code{data.frame} too.}

\item{window_size}{an \code{int} with the size of the sliding window.}

\item{matrix_profile}{multidimensional matrix profile (from \code{\link[=mstomp]{mstomp()}} or \code{\link[=mstomp_par]{mstomp_par()}}).}

\item{profile_index}{multidimensional profile index (from \code{\link[=mstomp]{mstomp()}} or \code{\link[=mstomp_par]{mstomp_par()}}).}

\item{n_bit}{an \code{int}. Number of bits for MDL discretization. (Default is \code{4}).}

\item{k}{an \code{int}. The number of MOTIFs to retrieve. \code{Inf} means all possible MOTIFs. (Default is
\code{Inf}).}
}
\value{
Returns the \code{motif_idx} with the index of MOTIFs founded and \code{motif_dim} with the spanned
dimensions of respective MOTIF.

Returns the \code{motif_idx} with the index of MOTIFs founded and \code{motif_dim} with the spanned
dimensions of respective MOTIF.
}
\description{
MDL Based MOTIF Discovery for Multidimensional Matrix Profile

MDL Based MOTIF Discovery for Multidimensional Matrix Profile
}
\details{
Although this functions handles Multivariate Time Series, it can also be used to handle
Univariate Time Series.

Although this functions handles Multivariate Time Series, it can also be used to handle
Univariate Time Series.
}
\examples{
# This is a fast toy example and results are useless. For a complete result, run the code inside
#'Not run' section below.
w <- mp_toy_data$sub_len
mp <- mstomp(mp_toy_data$data[1:200,], w, verbose = 0)
motifs <- m_motif_usearch(mp_toy_data$data[1:200,], w, mp$mp, mp$pi, 2)
\dontrun{
w <- mp_toy_data$sub_len
mp <- mstomp_par(mp_toy_data$data, w)
motifs <- m_motif_usearch(mp_toy_data$data, w, mp$mp, mp$pi, 4, 2)
}

# This is a fast toy example and results are useless. For a complete result, run the code inside
#'Not run' section below.
w <- mp_toy_data$sub_len
mp <- mstomp(mp_toy_data$data[1:200,], w, verbose = 0)
motifs <- m_motif_usearch(mp_toy_data$data[1:200,], w, mp$mp, mp$pi, 2)
\dontrun{
w <- mp_toy_data$sub_len
mp <- mstomp_par(mp_toy_data$data, w)
motifs <- m_motif_usearch(mp_toy_data$data, w, mp$mp, mp$pi, 4, 2)
}

}
\references{
\itemize{
\item Yeh CM, Kavantzas N, Keogh E. Matrix Profile VI : Meaningful Multidimensional Motif
Discovery.
}

\itemize{
\item Zhu Y, Imamura M, Nikovski D, Keogh E. Matrix Profile VII: Time Series Chains: A
New Primitive for Time Series Data Mining. Knowl Inf Syst. 2018 Jun 2;1–27.
}

Website: \url{https://sites.google.com/view/mstamp/}

Website: \url{http://www.cs.ucr.edu/~eamonn/MatrixProfile.html}

\itemize{
\item Yeh CM, Kavantzas N, Keogh E. Matrix Profile VI : Meaningful Multidimensional Motif
Discovery.
}

\itemize{
\item Zhu Y, Imamura M, Nikovski D, Keogh E. Matrix Profile VII: Time Series Chains: A
New Primitive for Time Series Data Mining. Knowl Inf Syst. 2018 Jun 2;1–27.
}

Website: \url{https://sites.google.com/view/mstamp/}

Website: \url{http://www.cs.ucr.edu/~eamonn/MatrixProfile.html}
}
\seealso{
\code{\link[=mstomp]{mstomp()}}, \code{\link[=mstomp_par]{mstomp_par()}}, \code{\link[=m_motif_search]{m_motif_search()}}

\code{\link[=mstomp]{mstomp()}}, \code{\link[=mstomp_par]{mstomp_par()}}, \code{\link[=m_motif_search]{m_motif_search()}}
}
