% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_pre.R
\name{mass.pre}
\alias{mass.pre}
\title{Precomputes several values used on MASS}
\usage{
mass.pre(data, data.size, query = NULL, query.size = NULL, window.size)
}
\arguments{
\item{data}{reference Time Series}

\item{data.size}{reference Time Series size}

\item{query}{query Time Series (default is NULL)}

\item{query.size}{query Time Series (default is NULL)}

\item{window.size}{sliding window size}
}
\value{
Returns data.fft, data.mean, data.sd, query.mean and query.sd
}
\description{
Precomputes several values used on MASS
}
\examples{
\dontrun{
w <- 30
d.size <- length(ref.data)
q.size <- length(query.data)

pre <- mass.pre(ref.data, d.size, query.data, q.size, w)

for(i in 1:(d.size - w + 1)) {
  dp <- mass(pre$data.fft, query.data[i:(i-1+w)], d.size, w, pre$data.mean, pre$data.sd,
          pre$query.mean[i], pre$query.sd[i])
}
}
}
\references{
Abdullah Mueen, Yan Zhu, Michael Yeh, Kaveh Kamgar, Krishnamurthy Viswanathan, Chetan Kumar Gupta and Eamonn Keogh (2015), The Fastest Similarity Search Algorithm for Time Series Subsequences under Euclidean Distance

\url{https://www.cs.unm.edu/~mueen/FastestSimilaritySearch.html}
}
\seealso{
\code{\link[=mass]{mass()}} for using precomputed values
}
