% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass.R
\name{mass}
\alias{mass}
\title{Calculates the distance profile using MASS algorithm}
\usage{
mass(data.fft, query.window, data.size, window.size, data.mean, data.sd,
  query.mean, query.sd)
}
\arguments{
\item{data.fft}{precomputed data product}

\item{query.window}{query window vector}

\item{data.size}{length of data}

\item{window.size}{length of query moving window}

\item{data.mean}{precomputed data moving average}

\item{data.sd}{precomputed data moving standard deviation}

\item{query.mean}{precomputed query average}

\item{query.sd}{precomputed query standard deviation}
}
\value{
Returns the distance profile for the given query and the last computed product
}
\description{
Mueen's Algorithm for Similarity Search is The Fastest Similarity Search Algorithm for Time Series Subsequences under Euclidean Distance and Correlation Coefficient.
}
\examples{
\dontrun{
w <- 30
d.size <- length(ref.data)
q.size <- length(query.data)

pre <- mass.pre(ref.data, d.size, query.data, q.size, w)

for(i in 1:(d.size - w + 1)) {
  dp <- mass(pre$data.fft, query.data[i:(i-1+w)], d.size, w, pre$data.mean, pre$data.sd,
          pre$query.mean[i], pre$query.sd[i])
}
}
}
\references{
Abdullah Mueen, Yan Zhu, Michael Yeh, Kaveh Kamgar, Krishnamurthy Viswanathan, Chetan Kumar Gupta and Eamonn Keogh (2015), The Fastest Similarity Search Algorithm for Time Series Subsequences under Euclidean Distance

\url{https://www.cs.unm.edu/~mueen/FastestSimilaritySearch.html}
}
\seealso{
\code{\link[=mass.pre]{mass.pre()}} to precomputation of input values
}
