% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple.R
\name{mass.simple.pre}
\alias{mass.simple.pre}
\title{Precomputes several values used on MASS}
\usage{
mass.simple.pre(data, data.size, window.size)
}
\arguments{
\item{data}{a \code{matrix} of \code{numeric}. Reference Time Series.}

\item{data.size}{an \code{int}. Reference Time Series size.}

\item{window.size}{an \code{int}. Sliding window size.}
}
\value{
Returns \code{data.fft} and \code{sumx2}.
}
\description{
The difference of this function to \code{\link[=mass.pre]{mass.pre()}} is that this does not normalize data. Specific for this domain.
}
\references{
\itemize{
\item Abdullah Mueen, Yan Zhu, Michael Yeh, Kaveh Kamgar, Krishnamurthy Viswanathan, Chetan Kumar Gupta and Eamonn Keogh (2015), The Fastest Similarity Search Algorithm for Time Series Subsequences under Euclidean Distance.
}

Website: \url{https://www.cs.unm.edu/~mueen/FastestSimilaritySearch.html}
}
\keyword{internal}
