% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple.R
\name{mass.simple}
\alias{mass.simple}
\title{Calculates the distance profile using MASS algorithm}
\usage{
mass.simple(data.fft, query.window, data.size, window.size, sumx2)
}
\arguments{
\item{data.fft}{precomputed data product.}

\item{query.window}{a \code{matrix} of \code{numeric}. Query window.}

\item{data.size}{an \code{int}. The length of the reference data.}

\item{window.size}{an \code{int}. Sliding window size.}

\item{sumx2}{precomputed sum of squares}
}
\value{
Returns the \code{distance.profile} for the given query and the \code{last.product} for STOMP algorithm and \code{sumy2}.
}
\description{
Mueen's Algorithm for Similarity Search is The Fastest Similarity Search Algorithm for Time Series Subsequences under Euclidean Distance and Correlation Coefficient.
The difference of this function to \code{\link[=mass]{mass()}} is that this does not normalize data. Specific for this domain.
}
\references{
Abdullah Mueen, Yan Zhu, Michael Yeh, Kaveh Kamgar, Krishnamurthy Viswanathan, Chetan Kumar Gupta and Eamonn Keogh (2015), The Fastest Similarity Search Algorithm for Time Series Subsequences under Euclidean Distance

\url{https://www.cs.unm.edu/~mueen/FastestSimilaritySearch.html}
}
\keyword{internal}
