% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skimp.R
\name{maximum_subsequence}
\alias{maximum_subsequence}
\title{Pan Matrix Profile maximum subsequence}
\usage{
maximum_subsequence(
  data,
  threshold = getOption("tsmp.pmp_ub", 0.95),
  refine_stepsize = getOption("tsmp.pmp_refine", 0.5),
  return_pmp = TRUE
)
}
\arguments{
\item{return_pmp}{}
}
\value{

}
\description{
Finds the upper bound
}
\examples{

}
