% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs.R
\name{motifs}
\alias{motifs}
\title{Search for Motifs}
\usage{
motifs(.mp, ...)
}
\arguments{
\item{.mp}{a TSMP object of class \code{MatrixProfile} or \code{MultiMatrixProfile}.}

\item{\dots}{further arguments to be passed to class specific function.}
}
\description{
Search for Motifs
}
