% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs-discords-small.R
\docType{data}
\name{motifs_discords_small}
\alias{motifs_discords_small}
\title{Just a synthetic dataset for testing}
\format{
A \code{vector} with 875 observations
}
\usage{
motifs_discords_small
}
\description{
Just a synthetic dataset for testing
}
\keyword{datasets}
