% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp-gait-data.R
\docType{data}
\name{mp_gait_data}
\alias{mp_gait_data}
\title{Original data used in the Time Series Chain demo}
\format{
A \code{matrix} with 904 rows and 1 column with the Y data from an accelerometer
}
\source{
\url{https://sites.google.com/site/timeserieschain/}
}
\usage{
mp_gait_data
}
\description{
Original data used in the Time Series Chain demo
}
\references{
\itemize{
\item Zhu Y, Imamura M, Nikovski D, Keogh E. Matrix Profile VII: Time Series Chains: A
New Primitive for Time Series Data Mining. Knowl Inf Syst. 2018 Jun 2;1-27.
}
}
\keyword{datasets}
