% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpx.R
\name{mpx}
\alias{mpx}
\title{Fast implementation of MP and MPI for internal purposes, without FFT}
\usage{
mpx(
  data,
  window_size,
  query = NULL,
  idx = TRUE,
  dist = c("euclidean", "pearson"),
  n_workers = 1
)
}
\arguments{
\item{data}{data}

\item{window_size}{window size}

\item{query}{query}

\item{idx}{calc and return indexes?}

\item{dist}{distance measure, euclidean or pearson?}

\item{n_workers}{theads for multi-threading}
}
\value{
Returns MP and MPI
}
\description{
Fast implementation of MP and MPI for internal purposes, without FFT
}
