% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstomp.R
\name{mstomp}
\alias{mstomp}
\title{Multivariate STOMP algorithm}
\usage{
mstomp(data, query.size, must.dim = NULL, exc.dim = NULL)
}
\arguments{
\item{data}{a matrix, where each colums is a time series (dimention). Can accept Lists and data.frames too.}

\item{query.size}{size of the sliding window}

\item{must.dim}{which dimentions to forcibly include (default is NULL)}

\item{exc.dim}{which dimentions to exclude (default is NULL)}
}
\value{
The matrix profile and profile index
}
\description{
Multivariate STOMP algorithm
}
\examples{

\dontrun{
mp <- mstomp(data, 30)
mp <- mstomp(data, 30, must.dim = c(1, 2))
mp <- mstomp(data, 30, exc.dim = c(2,3))
}
}
\references{
Yeh CM, Kavantzas N, Keogh E. Matrix Profile VI : Meaningful Multidimensional Motif Discovery.
}
\seealso{
\code{\link[=stamp]{stamp()}}
}
