% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{\%>\%}
\alias{\%>\%}
\alias{\%T>\%}
\alias{\%||\%}
\title{Pipe operator}
\usage{
lhs \%>\% rhs

lhs \%T>\% rhs

x \%||\% y
}
\description{
See \code{magrittr::\link[magrittr:pipe]{\%>\%}} for details.

See \code{magrittr::\link[magrittr:pipe]{\%T>\%}} for details.

See \code{rlang::\link[rlang:op-null-default]{\%||\%}} for details.
}
\keyword{internal}
