% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds.R
\name{perf}
\alias{perf}
\title{Future function to check performance}
\usage{
perf(truth, subs, window = 0)
}
\arguments{
\item{window}{}
}
\value{
Returns X,Y values for plotting
}
\description{
Future function to check performance
}
\examples{
perf(carfull$lab, subs)
perf(carsub$labIdx, subssub, carsub$subLen)
}
