% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{\%>\%}
\alias{\%>\%}
\alias{\%T>\%}
\title{Pipe operator}
\usage{
lhs \%>\% rhs

lhs \%T>\% rhs
}
\description{
See \code{magrittr::\link[magrittr:pipe]{\%>\%}} for details.

See \code{magrittr::\link[magrittr:pipe]{\%T>\%}} for details.
}
\keyword{internal}
