% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.ArcCounts}
\alias{plot.ArcCounts}
\title{Title}
\usage{
\method{plot}{ArcCounts}(.mp, data, type = c("data", "matrix"),
  exclusion_zone = NULL, edge_limit = NULL,
  threshold = quantile(.mp$mp, 0.1), main = "Arcs Discover",
  xlab = "index", ylab = "distance", ...)
}
\description{
Title
}
