% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.Arcs}
\alias{plot.Arcs}
\title{Title}
\usage{
\method{plot}{Arcs}(.mp, pairs, alpha = NULL, quality = 30, lwd = 15,
  col = c("blue", "orange"), main = "Arc Plot", ylab = "",
  xlab = "Profile Index", ...)
}
\description{
Title
}
