% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.MatrixProfile}
\alias{plot.MatrixProfile}
\title{Title}
\usage{
\method{plot}{MatrixProfile}(.mp, ylab = "distance", xlab = "index",
  main = "Unidimensional Matrix Profile", ...)
}
\description{
Title
}
