% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.Motif}
\alias{plot.Motif}
\title{Title}
\usage{
\method{plot}{Motif}(.mp, data, type = c("data", "matrix"), ncol = 3,
  main = "MOTIF Discover", xlab = "index", ylab = "distance", ...)
}
\description{
Title
}
