% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.MultiMatrixProfile}
\alias{plot.MultiMatrixProfile}
\title{Title}
\usage{
\method{plot}{MultiMatrixProfile}(.mp, ylab = "distance",
  xlab = "index", main = "Multidimensional Matrix Profile", ...)
}
\description{
Title
}
