% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.MultiMotif}
\alias{plot.MultiMotif}
\title{Title}
\usage{
\method{plot}{MultiMotif}(.mp, data, type = c("data", "matrix"),
  ncol = 3, main = "Multidimensional MOTIF Discover", xlab = "index",
  ylab = "distance", ...)
}
\description{
Title
}
