% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcplot.R, R/plot_arc.R
\name{plot_arc}
\alias{plot_arc}
\title{Plot arcs between indexes of a Profile Index}
\usage{
plot_arc(mp, pi, window_size, exclusion_zone = 5, edge_limit = 5,
  threshold = quantile(mp, 0.1), pairs = NULL, alpha = NULL,
  quality = 30, lwd = 15, col = c("blue", "orange"),
  main = "Arc Plot", ylab = "", xlab = "Profile Index", ...)

plot_arc(mp, pi, window_size, exclusion_zone = 5, edge_limit = 5,
  threshold = quantile(mp, 0.1), pairs = NULL, alpha = NULL,
  quality = 30, lwd = 15, col = c("blue", "orange"),
  main = "Arc Plot", ylab = "", xlab = "Profile Index", ...)
}
\arguments{
\item{mp}{a Matrix Profile, or a Corrected Arc Count profile from \code{\link[=fluss_cac]{fluss_cac()}}.}

\item{pi}{a Profile Index.}

\item{window_size}{an \code{int}. Size of the sliding window.}

\item{exclusion_zone}{an \code{int}. (Default is \code{5}). Exclusion zone for small distances. See
details.}

\item{edge_limit}{an \code{int}. (Default is \code{5}). Exclusion zone for the edges of the profile. See
details.}

\item{threshold}{a \code{numeric}. (Default is 10% of the lowest values). Threshold for retrieving
pairs. See details.}

\item{pairs}{a \code{matrix} with 2 columns. (Default is \code{NULL}). Instead of \code{mp}, \code{pi},
\code{window_size}, \code{edge_limit}, \code{threshold}, you can give your custom list of pairs.}

\item{alpha}{a \code{numeric}. (Default is \code{NULL}, automatic). Alpha value for lines transparency.}

\item{quality}{an \code{int}. (Default is \code{30}). Number of segments to draw the arc. Bigger value,
harder to render.}

\item{lwd}{an \code{int}. (Default is \code{15}). Line width.}

\item{col}{a \code{vector} of colors. (Default is \code{c("blue", "orange")}). Colors for right and left
arc, respectively. Accepts one color.}

\item{main}{a \code{string}. (Default is \code{"Arc Plot"}). Main title.}

\item{ylab}{a \code{string}. (Default is \code{""}). Y label}

\item{xlab}{a \code{string}. (Default is \code{"Profile Index"}). X label.}

\item{...}{further arguments to be passed to \code{\link[=plot]{plot()}}. See \code{\link[=par]{par()}}.}

\item{mp}{a Matrix Profile, or a Corrected Arc Count profile from \code{\link[=fluss_cac]{fluss_cac()}}.}

\item{pi}{a Profile Index.}

\item{window_size}{an \code{int}. Size of the sliding window.}

\item{exclusion_zone}{an \code{int}. (Default is \code{5}). Exclusion zone for small distances. See
details.}

\item{edge_limit}{an \code{int}. (Default is \code{5}). Exclusion zone for the edges of the profile. See
details.}

\item{threshold}{a \code{numeric}. (Default is 10% of the lowest values). Threshold for retrieving
pairs. See details.}

\item{pairs}{a \code{matrix} with 2 columns. (Default is \code{NULL}). Instead of \code{mp}, \code{pi},
\code{window_size}, \code{edge_limit}, \code{threshold}, you can give your custom list of pairs.}

\item{alpha}{a \code{numeric}. (Default is \code{NULL}, automatic). Alpha value for lines transparency.}

\item{quality}{an \code{int}. (Default is \code{30}). Number of segments to draw the arc. Bigger value,
harder to render.}

\item{lwd}{an \code{int}. (Default is \code{15}). Line width.}

\item{col}{a \code{vector} of colors. (Default is \code{c("blue", "orange")}). Colors for right and left
arc, respectively. Accepts one color.}

\item{main}{a \code{string}. (Default is \code{"Arc Plot"}). Main title.}

\item{ylab}{a \code{string}. (Default is \code{""}). Y label}

\item{xlab}{a \code{string}. (Default is \code{"Profile Index"}). X label.}

\item{...}{further arguments to be passed to \code{\link[=plot]{plot()}}. See \code{\link[=par]{par()}}.}
}
\value{
Quietly returns the computed \code{pairs}, or those you gave as input.

Quietly returns the computed \code{pairs}, or those you gave as input.
}
\description{
Sometimes may be useful to see where is the nearest neighbor graphically. This is the reasoning
behind, for example, FLUSS which uses the arc count to infer a semantic change, and SiMPle which
infer that arcs connect similar segments of a music. See details for a deeper explanation how to
use this function.

Sometimes may be useful to see where is the nearest neighbor graphically. This is the reasoning
behind, for example, FLUSS which uses the arc count to infer a semantic change, and SiMPle which
infer that arcs connect similar segments of a music. See details for a deeper explanation how to
use this function.
}
\details{
You have two options to use this function. First you can provide just the data, and the function
will try its best to retrieve the pairs for plotting. Second, you can skip the first parameters
and just provide the \code{pairs}, which is a \code{matrix} with two columns; the first is the starting
index, the second is the end index. Two colors are used to allow you to identify the direction of
the arc. If you use the \code{rpi} or \code{lpi} as input, you will see that these profile indexes have
just one direction.

\code{exclusion_zone} is used to filter out small arcs that may be useless (e.g. you may be interested
in similarities that are far away). \code{edge_limit} is used to filter out spurious arcs that are
used connect the beginning and the end of the profile (e.g. silent audio). \code{threshold} is used to
filter indexes that have distant nearest neighbor (e.g. retrieve only the best motifs).

You have two options to use this function. First you can provide just the data, and the function
will try its best to retrieve the pairs for plotting. Second, you can skip the first parameters
and just provide the \code{pairs}, which is a \code{matrix} with two columns; the first is the starting
index, the second is the end index. Two colors are used to allow you to identify the direction of
the arc. If you use the \code{rpi} or \code{lpi} as input, you will see that these profile indexes have
just one direction.

\code{exclusion_zone} is used to filter out small arcs that may be useless (e.g. you may be interested
in similarities that are far away). \code{edge_limit} is used to filter out spurious arcs that are
used connect the beginning and the end of the profile (e.g. silent audio). \code{threshold} is used to
filter indexes that have distant nearest neighbor (e.g. retrieve only the best motifs).
}
\examples{
plot_arc(pairs = matrix(c(5, 10, 1, 10, 20, 5), ncol = 2, byrow = TRUE))
plot_arc(pairs = matrix(c(5, 10, 1, 10, 20, 5), ncol = 2, byrow = TRUE))
}
