% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salient.R
\name{salient.score}
\alias{salient.score}
\title{Future function to check performance}
\usage{
salient.score(gtruth, subs, window = 0)
}
\arguments{
\item{gtruth}{Ground truth annotation.}

\item{subs}{Output from \code{salient.results}.}

\item{window}{Sliding window size.}
}
\value{
Returns X,Y values for plotting
}
\description{
Future function to check performance
}
\examples{
\dontrun{
  salient.score(carfull$lab, subs)
  salient.score(carsub$labIdx, subssub, carsub$subLen)
}

}
\keyword{internal}
