% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salient.R
\name{salient.subsequences}
\alias{salient.subsequences}
\title{Retrieve salient subsequences from a dataset}
\usage{
salient.subsequences(data, matrix.profile, profile.index, window.size,
  n.bits = 8, n.cand = 10, exclusion.zone = 1/2, verbose = 2)
}
\arguments{
\item{data}{a \code{vector}, column \code{matrix} or \code{data.frame}. If more than one column is provided, see
details.}

\item{matrix.profile}{a result from STAMP or STOMP algorithms.}

\item{profile.index}{a result from STAMP or STOMP algorithms.}

\item{window.size}{an \code{int} with the size of the sliding window.}

\item{n.bits}{an \code{int}. Number of bits for MDL discretization. (Default is \code{8}).}

\item{n.cand}{an \code{int}. number of candidate when picking the subsequence in each iteration.
(Default is \code{10}).}

\item{exclusion.zone}{a \code{numeric}. Size of the exclusion zone, based on \code{window.size}. (Default
is \code{1/2}). See details.}
}
\value{
Returns a \code{list} with \code{indexes}, a \code{vector} with the starting position of each
subsequence, \code{idx.bit.size}, a \code{vector} with the associated bitsize for each iteration and
\code{bits} the value used as input on \code{n.bits}.
}
\description{
In order to allow a meaningful visualization in Multi-Dimensional Space (MDS), this function
retrieves the most relevant subsequences using Minimal Description Length (MDL) framework.
}
\details{
The main purpose of this algorithm is to find subsequences in one time series, but this
implementation also covers the experimental effectiveness evaluation with "whole sequence"
setting. This means you can input a \code{matrix} where each column is a sequence and this algorithm
will retrieve the most relevant sequences. For this setting, the \code{exclusion.zone} is ignored, and
you need to pre-compute the ordinary euclidean distance matrix. See examples.

The \code{exclusion.zone} is used to avoid trivial matches.
}
\examples{
# subsequences setting (main purpose)
salient.subsequences(data, data$mp, data$pi, 30, n.bits = 8, n.cand = 10)

# sequences setting
dist.matrix <- as.matrix(dist(carfull$data))
mp <- matrix(0, nrow(carfull$data), 1)
pi <- matrix(0, nrow(carfull$data), 1)

for (i in 1:nrow(carfull$data)) {
  dist.matrix[i, i] <- Inf;
  pi[i] <- which.min(dist.matrix[i, ])
  mp[i] <- dist.matrix[i, pi[i]]
}

n.bits <- 8

subs <- salient.subsequences(t(carfull$data), mp, pi, 577, n.bits = n.bits, n.cand = 10)
cutoff <- which(diff(subs$idx.bit.size) > 0)[1] - 1
if(cutoff > 0) {
  carfull$lab[subs$indexes[1:(cutoff - 1)]]
} else {
  message("nothing to do")
}

}
\references{
\itemize{
\item Yeh CCM, Van Herle H, Keogh E. Matrix profile III: The matrix profile allows
visualization of salient subsequences in massive time series. Proc - IEEE Int Conf Data Mining,
ICDM. 2017;579–88.
}

\itemize{
\item Hu B, Rakthanmanon T, Hao Y, Evans S, Lonardi S, Keogh E. Discovering the Intrinsic
Cardinality and Dimensionality of Time Series Using MDL. In: 2011 IEEE 11th International
Conference on Data Mining. IEEE; 2011. p. 1086–91.
}

Website: \url{https://sites.google.com/site/salientsubs/}
}
