% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrimp.R
\name{scrimp}
\alias{scrimp}
\title{Anytime univariate SCRIMP algorithm (experimental)}
\usage{
scrimp(..., window.size, exclusion.zone = 1/2, s.size = Inf,
  verbose = 2)
}
\arguments{
\item{...}{a \code{matrix} or a \code{vector}.}

\item{window.size}{an \code{int}. Size of the sliding window.}

\item{exclusion.zone}{a \code{numeric}. Size of the exclusion zone, based on window size (default is
\code{1/2}). See details.}

\item{s.size}{a \code{numeric}. for anytime algorithm, represents the size (in observations) the
random calculation will occur (default is \code{Inf}).}

\item{verbose}{an \code{int}. See details. (Default is \code{2}).}
}
\value{
Returns the matrix profile \code{mp} and profile index \code{pi}.
}
\description{
Computes the best so far Matrix Profile and Profile Index for Univariate Time Series.
DISCLAIMER: This algorithm still in development by its authors.
Join similarity, RMP and LMP not implemented yet.
}
\details{
The Matrix Profile, has the potential to revolutionize time series data mining because of its
generality, versatility, simplicity and scalability. In particular it has implications for time
series motif discovery, time series joins, shapelet discovery (classification), density
estimation, semantic segmentation, visualization, rule discovery, clustering etc. The anytime
SCRIMP computes the Matrix Profile and Profile Index in such manner that it can be stopped before
its complete calculation and return the best so far results allowing ultra-fast approximate
solutions. \code{verbose} changes how much information is printed by this function; \code{0} means nothing,
\code{1} means text, \code{2} means text and sound. \code{exclusion.zone} is used to avoid  trivial matches.
}
\examples{
mp <- scrimp(toy_data$data[1:200,1], window.size = 30, verbose = 0)
\dontrun{
ref.data <- toy_data$data[,1]
query.data <- toy_data$data[,2]
# self similarity
mp <- scrimp(ref.data, window.size = 30, s.size = round(nrow(ref.data) * 0.1))
# join similarity
mp <- scrimp(ref.data, query.data, window.size = 30, s.size = round(nrow(query.data) * 0.1))
}
}
\references{
Website: \url{http://www.cs.ucr.edu/~eamonn/MatrixProfile.html}
}
\seealso{
\code{\link[=mstomp]{mstomp()}}, \code{\link[=mstomp.par]{mstomp.par()}}

Other matrix profile computations: \code{\link{mstomp.par}},
  \code{\link{mstomp}}, \code{\link{stamp.par}},
  \code{\link{stamp}}, \code{\link{stomp.par}},
  \code{\link{stomp}}
}
\concept{matrix profile computations}
