% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdts_train.R
\name{sdts.train}
\alias{sdts.train}
\title{Title}
\usage{
sdts.train(data, label, window.size, beta, pat.max, parallel = TRUE)
}
\arguments{
\item{data}{vector of doubles}

\item{label}{vector of bool annotation}

\item{window.size}{int or a vector of ints with window sizes}

\item{beta}{double}

\item{pat.max}{int or Inf, max number of shape feature captured}

\item{parallel}{use parallel computation inside}
}
\description{
Title
}
\examples{

}
