% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdts_predict.R
\name{sdts_predict}
\alias{sdts_predict}
\title{Scalable Dictionary learning for Time Series (SDTS) prediction function}
\usage{
sdts_predict(model, data, window_size)
}
\arguments{
\item{model}{a model created by SDTS training function \code{\link[=sdts_train]{sdts_train()}}.}

\item{data}{a \code{vector} of \code{numeric}. Time series.}

\item{window_size}{an \code{int}. The average sliding window size.}
}
\value{
Returns a \code{vector} of \code{logical} with predicted annotations.
}
\description{
Scalable Dictionary learning for Time Series (SDTS) prediction function
}
\examples{
# This is a fast toy example and results are useless. For a complete result, run the code inside
#'Not run' section below.
w <- c(110, 220)
subs <- 11000:20000
tr_data <- mp_test_data$train$data[subs]
tr_label <- mp_test_data$train$label[subs]
te_data <- mp_test_data$test$data[subs]
te_label <- mp_test_data$test$label[subs]
model <- sdts_train(tr_data, tr_label, w, verbose = 0)
predict <- sdts_predict(model, te_data, round(mean(w)))
sdts_score(te_label, predict, 1)
\dontrun{
windows <- c(110, 220, 330)
model <- sdts_train(mp_test_data$train$data, mp_test_data$train$label, windows, verbose = 0)
predict <- sdts_predict(model, mp_test_data$test$data, round(mean(windows)))
sdts_score(mp_test_data$test$label, predict, 1)
}
}
\references{
\itemize{
\item Yeh C-CM, Kavantzas N, Keogh E. Matrix profile IV: Using Weakly Labeled Time Series
to Predict Outcomes. Proc VLDB Endow. 2017 Aug 1;10(12):1802–12.
}

Website: \url{https://sites.google.com/view/weaklylabeled}
}
\seealso{
Other SDTS: \code{\link{sdts_score}},
  \code{\link{sdts_train}}
}
\concept{SDTS}
