% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{set_data}
\alias{set_data}
\title{Set/changes the data included in TSMP object.}
\usage{
set_data(.mp, data)
}
\arguments{
\item{.mp}{a TSMP object.}

\item{data}{a \code{matrix} (for one series) or a \code{list} of matrices (for two series).}
}
\value{
Returns silently the original TSMP object with changed data.
}
\description{
This may be useful if you want to include the data lately or remove the included data (set as \code{NULL}).
}
\examples{
mp <- tsmp(mp_toy_data$data[1:200, 1], window_size = 30, verbose = 0)
mp <- set_data(mp, NULL)
}
