% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple.R
\name{simple.fast}
\alias{simple.fast}
\title{Compute the similarity join for Sound data.}
\usage{
simple.fast(data, window.size, exclusion.zone = 1/2, verbose = 2)
}
\arguments{
\item{data}{a \code{matrix} of \code{numeric}, where each colums is a time series. Accepts \code{list} and \code{data.frame} too.}

\item{window.size}{an \code{int} with the size of the sliding window.}

\item{exclusion.zone}{a \code{numeric}. Size of the exclusion zone, based on query size (default is \code{1/2}).}

\item{verbose}{an \code{int}. See details. (Default is \code{2}).}
}
\value{
Returns a list with the Matrix Profile \code{mp} and Profile Index \code{pi}.
}
\description{
Compute the similarity join for Sound data.
}
\details{
\code{verbose} changes how much information is printed by this function; \code{0} means nothing, \code{1} means text, \code{2} means text and sound.
}
\examples{
w <- 30
data <- toy_data$data # 3 dimensions matrix
result <- simple.fast(data, w, verbose = 0)

}
\references{
\enumerate{
\item Silva D, Yeh C, Batista G, Keogh E. Simple: Assessing Music Similarity Using Subsequences Joins. Proc 17th ISMIR Conf. 2016;23–30.
}

\enumerate{
\item Silva DF, Yeh C-CM, Zhu Y, Batista G, Keogh E. Fast Similarity Matrix Profile for Music Analysis and Exploration. IEEE Trans Multimed. 2018;14(8):1–1.
}

Website: \url{https://sites.google.com/view/simple-fast}

Website: \url{https://sites.google.com/site/ismir2016simple/home}
}
