% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snippet.R
\name{snippetfinder}
\alias{snippetfinder}
\title{Time Series Snippets: A New Primitive for Time Series Data Mining}
\usage{
snippetfinder(data, N, sub, per)
}
\arguments{
\item{data}{Time Series}

\item{N}{number of snippets the user wishes to find}

\item{sub}{the length of snippet}

\item{per}{the MPdist subsequence length percentage. For example, if per is 100 then the MPdist subsequence is the same as sub.}
}
\value{
Returns the ## snippet : a list of N snippets
fraction : fraction of each snippet
snippetidx : the location of each
snippet within time sereis
}
\description{
MPdist is a recently introduced distance measure which considers two time series to be similar
}
\details{
MPdist
}
\examples{

}
\references{
\itemize{
\item Imani S, Madrid F, Ding W, Crouter S, Keogh E. Matrix Profile XIII:
Time Series Snippets: A New Primitive for Time Series Data Mining. In: 2018 IEEE International
Conference on Data Mining (ICDM). 2018.
}

\itemize{
\item Gharghabi S, Imani S, Bagnall A, Darvishzadeh A, Keogh E. Matrix Profile XII:
MPdist: A Novel Time Series Distance Measure to Allow Data Mining in More Challenging Scenarios.
In: 2018 IEEE International Conference on Data Mining (ICDM). 2018.
}

Website: \url{https://sites.google.com/site/snippetfinder/}
}
