% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp.R
\name{stamp}
\alias{stamp}
\title{Anytime univariate STAMP algorithm}
\usage{
stamp(..., window.size, exclusion.zone = 1/2, s.size = Inf)
}
\arguments{
\item{...}{a matrix or a vector, if a second time series is supplied it will be a join matrix profile}

\item{window.size}{size of the sliding window}

\item{exclusion.zone}{size of the exclusion zone, based on query size (default 1/2)}

\item{s.size}{for anytime algorithm, represents the size (in observations) the random calculation will occour (default Inf)}
}
\value{
The matrix profile and profile index
}
\description{
No description
}
\details{
No details
}
\examples{
\dontrun{
mp <- stamp(data, window.size = 30)
mp <- stamp(ref.data, query.data, window.size = 30, s.size = round(nrows(ref.data) * 0.1)) # 10 percent of data
}
}
\references{
Yeh CCM, Zhu Y, Ulanova L, Begum N, Ding Y, Dau HA, et al. Matrix profile I: All pairs similarity joins for time series: A unifying view that includes motifs, discords and shapelets. Proc - IEEE Int Conf Data Mining, ICDM. 2017;1317–22.

\url{http://www.cs.ucr.edu/~eamonn/MatrixProfile.html}
}
\seealso{
\code{\link[=mstomp]{mstomp()}}

Other Stamp: \code{\link{stamp.par}}
}
\concept{Stamp}
