% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stampi.R
\name{stampi_update}
\alias{stampi_update}
\title{Real-time STAMP algorithm}
\usage{
stampi_update(.mp, new_data)
}
\arguments{
\item{.mp}{a TSMP object of class \code{MatrixProfile}.}

\item{new_data}{new data to append to original data.}
}
\value{
Returns the input \code{.mp} updated with the new information.
}
\description{
Real-time STAMP algorithm
}
