% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{std}
\alias{std}
\title{Population SD, as R always calculate with n-1 (sample), here we fix it}
\usage{
std(data)
}
\arguments{
\item{data}{a \code{vector} or a column \code{matrix} of \code{numeric}.}
}
\value{
Returns the corrected standard deviation from sample to population
}
\description{
Population SD, as R always calculate with n-1 (sample), here we fix it
}
\keyword{internal}
