% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stampi.R
\name{stompi_update}
\alias{stompi_update}
\title{Real-time STOMP algorithm}
\usage{
stompi_update(.mp, new_data)
}
\arguments{
\item{.mp}{a TSMP object of class \code{MatrixProfile}.}

\item{new_data}{new data to append to original data.}
}
\value{
Returns the input \code{.mp} updated with the new information.
}
\description{
Real-time STOMP algorithm
}
