% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_data.R
\docType{data}
\name{toy_data}
\alias{toy_data}
\title{Original data used in the mSTAMP demo}
\format{A \code{list} with a \code{matrix} with 550 rows and 3 dimensions and an \code{int}: \describe{
\item{data}{data with embedded MOTIFs} \item{sub.len}{size of sliding window} }}
\source{
\url{https://sites.google.com/view/mstamp/}
}
\usage{
toy_data
}
\description{
A synthetic dataset with embedded MOTIFs for multidimensional discovery
}
\references{
\itemize{
\item Yeh CM, Kavantzas N, Keogh E. Matrix Profile VI : Meaningful Multidimensional Motif
Discovery.
}

Website: \url{http://www.cs.ucr.edu/~eamonn/MatrixProfile.html}
}
\keyword{datasets}
