% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsmp-package.R
\docType{package}
\name{tsmp}
\alias{tsmp}
\alias{tsmp-package}
\title{Time Series with Matrix Profile}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

A toolkit implementing the Matrix Profile concept that was created by CS-UCR <http://www.cs.ucr.edu/~eamonn/MatrixProfile.html>.
}
\references{
\url{http://www.cs.ucr.edu/~eamonn/MatrixProfile.html}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/franzbischoff/tsmp}
  \item Report bugs at \url{https://github.com/franzbischoff/tsmp/issues}
}

}
\author{
\strong{Maintainer}: Francisco Bischoff \email{fbischoff@med.up.pt} (https://orcid.org/0000-0002-5301-8672)

Other contributors:
\itemize{
  \item Michael Yeh \email{myeh003@ucr.edu} (https://orcid.org/0000-0002-9807-2963) [researcher, conceptor, contributor]
  \item Diego Silva \email{diegofs@ufscar.br} (https://orcid.org/0000-0002-5184-9413) [researcher, conceptor, contributor]
  \item Yan Zhu \email{yzhu015@ucr.edu} [researcher, conceptor, contributor]
}

}
