% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{zero.crossings}
\alias{zero.crossings}
\title{Counts number of zero-crossings}
\usage{
zero.crossings(data)
}
\arguments{
\item{data}{is the input time-domain signal (one dimensional).}
}
\value{
Returns the amount of zero-crossings in the input signal.
}
\description{
Count the number of zero-crossings from the supplied time-domain input vector. A simple method is
applied here that can be easily ported to a real-time system that would minimize the number of
if-else conditionals.
}
\references{
\url{https://www.dsprelated.com/showcode/179.php}
}
\author{
sparafucile17 06/27/04
}
\keyword{internal}
