% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{zero.one.norm}
\alias{zero.one.norm}
\title{Normalizes data between Zero and One}
\usage{
zero.one.norm(data)
}
\arguments{
\item{data}{a \code{vector} or a column \code{matrix} of \code{numeric}.}
}
\value{
Returns the normalized data.
}
\description{
Normalizes data between Zero and One
}
\keyword{internal}
