% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{znorm}
\alias{znorm}
\title{Normalizes data for mean Zero and Standard Deviation One}
\usage{
znorm(data)
}
\arguments{
\item{data}{a \code{vector} or a column \code{matrix} of \code{numeric}.}
}
\value{
Returns the normalized data
}
\description{
Normalizes data for mean Zero and Standard Deviation One
}
\keyword{internal}
