% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdts_train.R
\name{compute.f.meas}
\alias{compute.f.meas}
\title{Computes de F-Score}
\usage{
compute.f.meas(label, pos.st, pos.ed, dist.pro, thold, window.size, beta)
}
\arguments{
\item{label}{a vector with the data bool annotation}

\item{pos.st}{a vector with the starting points of label}

\item{pos.ed}{a vector with the ending points of label}

\item{dist.pro}{the distance profile of the data}

\item{thold}{a number with the threshold used to compute the prediction}

\item{window.size}{an integer with the sliding window size}

\item{beta}{a number that balance the F-Score. Beta > 1 towards recall, < towards precision}
}
\value{
Returns the F-Score, precision and recall values
}
\description{
Computes de F-Score
}
\keyword{internal}
