% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discords.R
\name{discords}
\alias{discords}
\title{Search for Discord}
\usage{
discords(.mp, ...)
}
\arguments{
\item{.mp}{a TSMP object of class \code{MatrixProfile}}

\item{\dots}{further arguments to be passed to class specific function.}
}
\description{
Search for Discord
}
