% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{fast.movavg}
\alias{fast.movavg}
\title{Fast implementation of moving average using filter}
\usage{
fast.movavg(data, n)
}
\arguments{
\item{data}{a vector}

\item{n}{moving average window}
}
\value{
Returns a vector with the moving average
}
\description{
Fast implementation of moving average using filter
}
\examples{
data.avg <- fast.movavg(toy_data$data[,1], toy_data$sub.len)
}
