% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{fast.movsd}
\alias{fast.movsd}
\title{Fast implementation of moving standard deviation using filter}
\usage{
fast.movsd(data, window.size)
}
\arguments{
\item{data}{a \code{vector} or a column \code{matrix} of \code{numeric}.}

\item{window.size}{moving sd window size}
}
\value{
Returns a \code{vector} with the moving standard deviation
}
\description{
Fast implementation of moving standard deviation using filter
}
\examples{
data.sd <- fast.movsd(toy_data$data[,1], toy_data$sub.len)
}
