% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_motifs.R
\name{find_motifs}
\alias{find_motifs}
\title{Find MOTIFs and Plot them}
\usage{
find_motifs(matrix_profile, profile_index, window_size, data,
  exclusion_zone = 1/2, radius = 3, n_motifs = 3, plot = TRUE,
  cols = 3)
}
\arguments{
\item{cols}{}
}
\description{
Find MOTIFs and Plot them
}
