% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluss.R
\name{floss}
\alias{floss}
\title{FLOSS}
\usage{
floss(.mp, new_data, num_segments, data_window, exclusion_zone = NULL,
  chunk_size = NULL)
}
\arguments{
\item{.mp}{.mp a TSMP object of class \code{MatrixProfile}.}

\item{data_window}{if a \code{number} will be used instead of embedded value. (Default is \code{NULL}).}

\item{exclusion_zone}{if a \code{number} will be used instead of embedded value. (Default is \code{NULL}).}
}
\description{
FLOSS
}
