% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluss.R
\name{fluss}
\alias{fluss}
\title{Fast Low-cost Unipotent Semantic Segmentation (FLUSS)}
\usage{
fluss(.mp, num_segments, exclusion_zone = NULL)
}
\arguments{
\item{.mp}{a TSMP object of class \code{ArcCount}.}

\item{num_segments}{an \code{int}. Number of segments to extract. Based on domain knowledge.}

\item{exclusion_zone}{if a \code{number} will be used instead of embeded value. (Default is \code{NULL}).}
}
\value{
Returns a list with \code{segments} (location of semantic changes), \code{mp} (matrix profile if
computed), \code{pi} (profile index, input of computed), \code{cac} corrected arc count.
}
\description{
FLUSS is a Domain Agnostic Online Semantic Segmentation that uses the assumption that when few
arc are crossing a given index point, means that there is a high probability of semantic change.
}
\examples{
# This is a fast toy example and results are useless. For a complete result, run the code inside
#'Not run' section below.
data <- mp_fluss_data$tilt_abp$data[1:1000]
truth <- 400
w <- 10
segments <- fluss(data, w, 1, gtruth = truth, verbose = 0)
\dontrun{
data <- mp_fluss_data$walkjogrun$data
w <- mp_fluss_data$walkjogrun$window # 80
truth <- mp_fluss_data$walkjogrun$gtruth # 3800 6800
nseg <- length(mp_fluss_data$walkjogrun$gtruth) # 2
segments <- fluss(data, w, nseg, gtruth = truth)
}
}
\references{
\itemize{
\item Gharghabi S, Ding Y, Yeh C-CM, Kamgar K, Ulanova L, Keogh E. Matrix Profile VIII:
Domain Agnostic Online Semantic Segmentation at Superhuman Performance Levels. In: 2017 IEEE
International Conference on Data Mining (ICDM). IEEE; 2017. p. 117–26.
}

Website: \url{https://sites.google.com/site/onlinesemanticsegmentation/}

Website: \url{http://www.cs.ucr.edu/~eamonn/MatrixProfile.html}
}
\seealso{
Other Semantic Segmentations: \code{\link{fluss_cac}},
  \code{\link{fluss_extract}}, \code{\link{fluss_score}}
}
\concept{Semantic Segmentations}
