% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluss.R
\name{fluss_cac}
\alias{fluss_cac}
\title{FLUSS - Corrected Arc Counts}
\usage{
fluss_cac(.mp, exclusion_zone = NULL)
}
\arguments{
\item{.mp}{}

\item{exclusion_zone}{a \code{numeric}. Size of the exclusion zone, based on window size (default is \code{5}).}
}
\value{
Returns a companion matrix with the same size of profile index. This matrix contains the number of
'arcs' crossing over each index.
}
\description{
Computes the arc count with edge correction (CAC).
}
\details{
Original paper suggest using the classic statistical-process-control heuristic to set a threshold
where a semantic change may occur in CAC. This may be useful in real-time implementation as we don't
know in advance the number of domain changes to look for. Please check original paper (1).
}
\examples{
# This is a fast toy example and results are useless. For a complete result, run the code inside
#'Not run' section below.
data <- mp_fluss_data$tilt_abp$data[1:1000]
w <- 210
mp <- stomp(data, window_size = w, verbose = 0)
cac <- fluss_cac(mp$pi, w)

\dontrun{
data <- mp_fluss_data$walkjogrun$data
w <- mp_fluss_data$walkjogrun$window # 80
mp <- stomp(data, window_size = w)
cac <- fluss_cac(mp$pi, w)
}
}
\references{
\itemize{
\item Gharghabi S, Ding Y, Yeh C-CM, Kamgar K, Ulanova L, Keogh E. Matrix Profile VIII: Domain Agnostic Online Semantic Segmentation at Superhuman Performance Levels. In: 2017 IEEE International Conference on Data Mining (ICDM). IEEE; 2017. p. 117–26.
}

Website: \url{https://sites.google.com/site/onlinesemanticsegmentation/}

Website: \url{http://www.cs.ucr.edu/~eamonn/MatrixProfile.html}
}
\seealso{
Other fluss: \code{\link{fluss_extract}},
  \code{\link{fluss_score}}, \code{\link{fluss}}
}
\concept{fluss}
