% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluss.R
\name{fluss_extract}
\alias{fluss_extract}
\title{FLUSS - Extract Segments}
\usage{
fluss_extract(.mpac, num_segments, exclusion_zone = NULL)
}
\arguments{
\item{num_segments}{an \code{int}. Number of segments to extract. Based on domain knowledge.}

\item{exclusion_zone}{a \code{numeric}. Size of the exclusion zone, based on window size (default is
\code{5}).}

\item{arc_counts}{a \code{matrix} with the corrected arc counts from \code{\link[=fluss_cac]{fluss_cac()}}.}

\item{window_size}{an \code{int}. Size of the sliding window.}
}
\value{
Returns an \code{int} or a \code{vector} of \code{int} with the location of predicted semantic changes.
The number of locations is not greater than \code{num_segments}.
}
\description{
Extract candidate points of semantic changes.
}
\examples{
# This is a fast toy example and results are useless. For a complete result, run the code inside
#'Not run' section below.
data <- mp_fluss_data$tilt_abp$data[1:1000]
w <- 210
mp <- stomp(data, window_size = w, verbose = 0)
cac <- fluss_cac(mp$pi, w)
segments <- fluss_extract(cac, 1, w)
\dontrun{
data <- mp_fluss_data$walkjogrun$data
w <- mp_fluss_data$walkjogrun$window # 80
nseg <- length(mp_fluss_data$walkjogrun$gtruth) # 2
mp <- stomp(data, window_size = w)
cac <- fluss_cac(mp$pi, w)
segments <- fluss_extract(cac, nseg, w)
}
}
\references{
\itemize{
\item Gharghabi S, Ding Y, Yeh C-CM, Kamgar K, Ulanova L, Keogh E. Matrix Profile VIII:
Domain Agnostic Online Semantic Segmentation at Superhuman Performance Levels. In: 2017 IEEE
International Conference on Data Mining (ICDM). IEEE; 2017. p. 117–26.
}

Website: \url{https://sites.google.com/site/onlinesemanticsegmentation/}

Website: \url{http://www.cs.ucr.edu/~eamonn/MatrixProfile.html}
}
\seealso{
Other fluss: \code{\link{fluss_cac}},
  \code{\link{fluss_score}}, \code{\link{fluss}}
}
\concept{fluss}
