% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdts_train.R
\name{golden.section.2}
\alias{golden.section.2}
\title{Title}
\usage{
golden.section.2(dist.pro, thold, label, pos.st, pos.ed, beta, window.size,
  fit.idx)
}
\arguments{
\item{fit.idx}{}
}
\description{
Title
}
\keyword{internal}
