% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdts_train.R
\name{golden.section}
\alias{golden.section}
\title{Title}
\usage{
golden.section(dist.pro, label, pos.st, pos.ed, beta, window.size)
}
\arguments{
\item{window.size}{}
}
\description{
Title
}
\keyword{internal}
