% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_pre.R
\name{mass.pre}
\alias{mass.pre}
\title{Precomputes several values used on MASS}
\usage{
mass.pre(data, data.size, query = NULL, query.size = NULL, window.size)
}
\arguments{
\item{data}{reference Time Series}

\item{data.size}{reference Time Series size}

\item{query}{query Time Series (default is NULL)}

\item{query.size}{query Time Series (default is NULL)}

\item{window.size}{sliding window size}
}
\value{
Returns data.fft, data.mean, data.sd, query.mean and query.sd
}
\description{
Precomputes several values used on MASS
}
\references{
Abdullah Mueen, Yan Zhu, Michael Yeh, Kaveh Kamgar, Krishnamurthy Viswanathan, Chetan Kumar Gupta and Eamonn Keogh (2015), The Fastest Similarity Search Algorithm for Time Series Subsequences under Euclidean Distance

\url{https://www.cs.unm.edu/~mueen/FastestSimilaritySearch.html}
}
