% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstomp_par.R
\name{mstomp.par}
\alias{mstomp.par}
\title{Multivariate STOMP algorithm Parallel version}
\usage{
mstomp.par(data, window.size, exclusion.zone = 1/2)
}
\arguments{
\item{data}{a \code{matrix} of \code{numeric}, where each colums is a time series. Accepts \code{vector} (see details), \code{list} and \code{data.frame} too.}

\item{window.size}{an \code{int}. Size of the sliding window.}

\item{exclusion.zone}{an \code{int}. Size of the exclusion zone, based on query size (default is \code{1/2}).}
}
\value{
Returns the matrix profile \code{mp} and profile index \code{pi}.
It also returns the left and right matrix profile \code{lmp}, \code{rmp} and profile index \code{lpi}, \code{rpi} that may be used to detect Time Series Chains (Yan Zhu 2018).
}
\description{
Computes the Matrix Profile and Profile Index for Multivariate Time Series.
}
\details{
The Matrix Profile, has the potential to revolutionize time series data mining because of its generality, versatility, simplicity and scalability. In particular it has implications for time series motif discovery, time series joins, shapelet discovery (classification), density estimation, semantic segmentation, visualization, rule discovery, clustering etc.
The MSTOMP computes the Matrix Profile and Profile Index for Multivariate Time Series that is meaningful for multidimensional MOTIF discovery. It uses the STOMP algorithm that is faster than STAMP but lacks its anytime property.

Although this functions handles Multivariate Time Series, it can also be used to handle Univariate Time Series.
}
\examples{
\dontrun{
mp <- mstomp.par(data, 30)
}
}
\references{
\enumerate{
\item Yeh CM, Kavantzas N, Keogh E. Matrix Profile VI : Meaningful Multidimensional Motif Discovery.
}

\enumerate{
\item Zhu Y, Imamura M, Nikovski D, Keogh E. Matrix Profile VII: Time Series Chains: A New Primitive for Time Series Data Mining. Knowl Inf Syst. 2018 Jun 2;1–27.
}

Website: \url{https://sites.google.com/view/mstamp/}

Website: \url{http://www.cs.ucr.edu/~eamonn/MatrixProfile.html}
}
\seealso{
\code{\link[=stamp.par]{stamp.par()}}

Other Stomp: \code{\link{mstomp}}
}
\concept{Stomp}
