% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstomp.R
\name{mstomp}
\alias{mstomp}
\title{Multivariate STOMP algorithm}
\usage{
mstomp(data, window.size, must.dim = NULL, exc.dim = NULL,
  exclusion.zone = 1/2)
}
\arguments{
\item{data}{a matrix, where each colums is a time series (dimention). Can accept Lists and data.frames too.}

\item{window.size}{size of the sliding window}

\item{must.dim}{which dimentions to forcibly include (default is NULL)}

\item{exc.dim}{which dimentions to exclude (default is NULL)}
}
\value{
The matrix profile and profile index
}
\description{
No description
}
\details{
No details
}
\examples{
\dontrun{
mp <- mstomp(data, 30)
mp <- mstomp(data, 30, must.dim = c(1, 2))
mp <- mstomp(data, 30, exc.dim = c(2,3))
}
}
\references{
Yeh CM, Kavantzas N, Keogh E. Matrix Profile VI : Meaningful Multidimensional Motif Discovery.
\link{https://sites.google.com/view/mstamp/}
}
\seealso{
\code{\link[=stamp]{stamp()}}
}
