% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{read}
\alias{read}
\title{Read object from file}
\usage{
read(x, ...)
}
\arguments{
\item{x}{a \code{character} string with the input filename.}

\item{...}{other arguments to be passed forward.}
}
\description{
Read object from file
}
\examples{

\dontrun{
result <- read("input.json")
}
}
