% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skimp.R
\name{skimp}
\alias{skimp}
\title{Title}
\usage{
skimp(data, window_sizes = seq.int(4, seq_end, jump), plot = TRUE)
}
\arguments{
\item{window_sizes}{The work closest in spirit to ours is VALMOD. The idea of VALMOD is to compute the MP for
the shortest length of interest, then use the information gleaned from it to guide a search
through longer subsequence lengths, exploiting lower bounds to prune off some calculations.
This idea works well for the first few of the longer subsequence lengths, but the lower bounds
progressively weaken, making the pruning ineffective. Thus, in the five case studies they
presented, the mean value of U/L was just 1.24. In contrast, consider that our termite example
in Fig. 15 has a U/L ratio of 240, more than two orders of magnitude larger. Thus, VALMOD is
perhaps best seen as finding motifs with some tolerance for a slightly (~25%) too short
user-specified query length, rather than a true "motif-of-all-lengths" algorithm. Also note
that apart from the shortest length, VALMOD only gives some information for the other lengths,
unlike pmp, which contains exact distances for all subsequences of all lengths.}

\item{plot}{plot each interaction}
}
\value{
Returns the Pan Matrix Profile
}
\description{
Title
}
