% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp.R, R/stamp_par.R
\name{stamp}
\alias{stamp}
\alias{stamp_par}
\title{Anytime univariate STAMP algorithm}
\usage{
stamp(..., window_size, exclusion_zone = 1/2, verbose = 2,
  s_size = Inf)

stamp_par(..., window_size, exclusion_zone = 1/2, verbose = 2,
  s_size = Inf, n_workers = 2)
}
\arguments{
\item{...}{a \code{matrix} or a \code{vector}. If a second time series is supplied it will be a join matrix
profile.}

\item{window_size}{an \code{int}. Size of the sliding window.}

\item{exclusion_zone}{a \code{numeric}. Size of the exclusion zone, based on window size (default is
\code{1/2}). See details.}

\item{verbose}{an \code{int}. See details. (Default is \code{2}).}

\item{s_size}{a \code{numeric}. for anytime algorithm, represents the size (in observations) the
random calculation will occur (default is \code{Inf}).}

\item{n_workers}{an \code{int}. Number of workers for parallel. (Default is \code{2}).}
}
\value{
Returns the matrix profile \code{mp} and profile index \code{pi}. It also returns the left and
right matrix profile \code{lmp}, \code{rmp} and profile index \code{lpi}, \code{rpi} that may be used to detect
Time Series Chains (Yan Zhu 2018).
}
\description{
Computes the best so far Matrix Profile and Profile Index for Univariate Time Series.
}
\details{
The Matrix Profile, has the potential to revolutionize time series data mining because of its
generality, versatility, simplicity and scalability. In particular it has implications for time
series motif discovery, time series joins, shapelet discovery (classification), density
estimation, semantic segmentation, visualization, rule discovery, clustering etc. The anytime
STAMP computes the Matrix Profile and Profile Index in such manner that it can be stopped before
its complete calculation and return the best so far results allowing ultra-fast approximate
solutions. \code{verbose} changes how much information is printed by this function; \code{0} means nothing,
\code{1} means text, \code{2} means text and sound. \code{exclusion_zone} is used to avoid  trivial matches; if
a query data is provided (join similarity), this parameter is ignored.
}
\section{Functions}{
\itemize{
\item \code{stamp}: Single thread version.

\item \code{stamp_par}: Parallel version.
}}

\examples{
mp <- stamp(mp_toy_data$data[1:200,1], window_size = 30, verbose = 0)
\dontrun{
ref_data <- mp_toy_data$data[,1]
query_data <- mp_toy_data$data[,2]
# self similarity
mp <- stamp(ref_data, window_size = 30, s_size = round(nrow(ref_data) * 0.1))
# join similarity
mp <- stamp(ref_data, query_data, window_size = 30, s_size = round(nrow(query_data) * 0.1))
}
}
\references{
\itemize{
\item Yeh CCM, Zhu Y, Ulanova L, Begum N, Ding Y, Dau HA, et al. Matrix profile I: All
pairs similarity joins for time series: A unifying view that includes motifs, discords and
shapelets. Proc - IEEE Int Conf Data Mining, ICDM. 2017;1317–22.
}

\itemize{
\item Zhu Y, Imamura M, Nikovski D, Keogh E. Matrix Profile VII: Time Series Chains: A
New Primitive for Time Series Data Mining. Knowl Inf Syst. 2018 Jun 2;1–27.
}

Website: \url{http://www.cs.ucr.edu/~eamonn/MatrixProfile.html}
}
\seealso{
Other matrix profile computations: \code{\link{mstomp}},
  \code{\link{scrimp}}, \code{\link{stomp}},
  \code{\link{tsmp}}
}
\concept{matrix profile computations}
