% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp.R
\name{stamp}
\alias{stamp}
\title{Anytime univariate STAMP algorithm}
\usage{
stamp(data, query.size, exclusion.zone = 1/2, s.size = Inf)
}
\arguments{
\item{data}{a matrix, with time series as a column}

\item{query.size}{size of the sliding window}

\item{exclusion.zone}{size of the exclusion zone, based on query size (default 1/2)}

\item{s.size}{for anytime algorithm, represents the size (in observations) the random calculation will occour (default Inf)}
}
\value{
The matrix profile and profile index
}
\description{
No description
}
\details{
No details
}
\examples{

\dontrun{
mp <- stamp(data, 30)
}
}
\references{
Yeh CCM, Zhu Y, Ulanova L, Begum N, Ding Y, Dau HA, et al. Matrix profile I: All pairs similarity joins for time series: A unifying view that includes motifs, discords and shapelets. Proc - IEEE Int Conf Data Mining, ICDM. 2017;1317–22.
}
\seealso{
\code{\link[=mstomp]{mstomp()}}
}
