% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{std}
\alias{std}
\title{Population SD, as R always calculate with n-1 (sample), here we fix it}
\usage{
std(x)
}
\arguments{
\item{x}{a vector}
}
\value{
Returns a corrected standard deviation from sample to population
}
\description{
Population SD, as R always calculate with n-1 (sample), here we fix it
}
\keyword{internal}
