% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsmp-package.R
\docType{package}
\name{tsmp-package}
\alias{tsmp}
\alias{tsmp-package}
\title{tsmp: Time Series with Matrix Profile}
\description{
A toolkit implementing the Matrix Profile concept that was created by CS-UCR <http://www.cs.ucr.edu/~eamonn/MatrixProfile.html>.
}
\details{
Details in progress
}
\references{
\url{http://www.cs.ucr.edu/~eamonn/MatrixProfile.html}
}
\author{
\strong{Maintainer}: Francisco Bischoff \email{fbischoff@med.up.pt}

Authors:
\itemize{
  \item Michael Yeh \email{myeh003@ucr.edu} [contributor]
}

}
\keyword{internal}
