% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsmp-package.R
\docType{package}
\name{tsmp-package}
\alias{tsmp}
\alias{tsmp-package}
\title{Time Series with Matrix Profile}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

A toolkit implementing the Matrix Profile concept that was created by CS-UCR <http://www.cs.ucr.edu/~eamonn/MatrixProfile.html>.
}
\references{
\itemize{
\item Abdullah Mueen, Yan Zhu, Michael Yeh, Kaveh Kamgar, Krishnamurthy Viswanathan,
Chetan Kumar Gupta and Eamonn Keogh (2015), The Fastest Similarity Search Algorithm for Time
Series Subsequences under Euclidean Distance
}

\itemize{
\item Dau HA, Keogh E. Matrix Profile V: A Generic Technique to Incorporate Domain
Knowledge into Motif Discovery. In: Proceedings of the 23rd ACM SIGKDD International Conference
on Knowledge Discovery and Data Mining - KDD ’17. New York, New York, USA: ACM Press; 2017. p.
125–34.
}

\itemize{
\item Gharghabi S, Ding Y, Yeh C-CM, Kamgar K, Ulanova L, Keogh E. Matrix Profile VIII:
Domain Agnostic Online Semantic Segmentation at Superhuman Performance Levels. In: 2017 IEEE
International Conference on Data Mining (ICDM). IEEE; 2017. p. 117–26.
}

\itemize{
\item Heldt, T., Oefinger, M.B., Hoshiyama, M. and Mark, R.G., 2003, September.
Circulatory response to passive and active changes in posture. In IEEE Computers in Cardiology,
2003 (pp. 263-266).
}

\itemize{
\item Hu B, Rakthanmanon T, Hao Y, Evans S, Lonardi S, Keogh E. Discovering the Intrinsic
Cardinality and Dimensionality of Time Series Using MDL. In: 2011 IEEE 11th International
Conference on Data Mining. IEEE; 2011. p. 1086–91.
}

\itemize{
\item Reiss, A. and Stricker, D., 2012. Introducing a new benchmarked dataset for
activity monitoring. In 16th International Symposium on Wearable Computers (ISWC), 2012, pages
108–109. IEEE, 2012.
}

\itemize{
\item Roverso, D., Multivariate temporal classification by windowed wavelet decomposition
and recurrent neural networks, in 3rd ANS Int’l Topical Meeting on Nuclear Plant
Instrumentation, Control and Human-Machine Interface, vol. 20, Washington, DC, USA, 2000.
}

\itemize{
\item Silva D, Yeh C, Batista G, Keogh E. Simple: Assessing Music Similarity Using
Subsequences Joins. Proc 17th ISMIR Conf. 2016;23–30.
}

\itemize{
\item Silva DF, Yeh C-CM, Zhu Y, Batista G, Keogh E. Fast Similarity Matrix Profile for
Music Analysis and Exploration. IEEE Trans Multimed. 2018;14(8):1–1.
}

\itemize{
\item Yeh CCM, Van Herle H, Keogh E. Matrix profile III: The matrix profile allows
visualization of salient subsequences in massive time series. Proc - IEEE Int Conf Data Mining,
ICDM. 2017;579–88.
}

\itemize{
\item Yeh C-CM, Kavantzas N, Keogh E. Matrix profile IV: Using Weakly Labeled Time Series
to Predict Outcomes. Proc VLDB Endow. 2017 Aug 1;10(12):1802–12.
}

\itemize{
\item Yeh CM, Kavantzas N, Keogh E. Matrix Profile VI : Meaningful Multidimensional Motif
Discovery.
}

\itemize{
\item Zhu Y, Imamura M, Nikovski D, Keogh E. Matrix Profile VII: Time Series Chains: A
New Primitive for Time Series Data Mining. Knowl Inf Syst. 2018 Jun 2;1–27.
}

Website: \url{http://www.cs.ucr.edu/~eamonn/MatrixProfile.html}

Website: \url{https://www.cs.unm.edu/~mueen/FastestSimilaritySearch.html}

Website: \url{https://sites.google.com/site/timeserieschain/}

Website: \url{https://sites.google.com/site/onlinesemanticsegmentation/}

Website: \url{https://sites.google.com/site/salientsubs/}

Website: \url{https://sites.google.com/view/mstamp/}

Website: \url{https://sites.google.com/view/simple-fast/}

Website: \url{https://sites.google.com/site/ismir2016simple/home/}

Website: \url{https://sites.google.com/view/weaklylabeled/}

Website: \url{http://www.cs.ucr.edu/~hdau001/guided_motif_search/}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/franzbischoff/tsmp}
  \item Report bugs at \url{https://github.com/franzbischoff/tsmp/issues}
}

}
\author{
\strong{Maintainer}: Francisco Bischoff \email{fbischoff@med.up.pt} (https://orcid.org/0000-0002-5301-8672)

Other contributors:
\itemize{
  \item Michael Yeh \email{myeh003@ucr.edu} (https://orcid.org/0000-0002-9807-2963) [researcher, conceptor, contributor]
  \item Diego Silva \email{diegofs@ufscar.br} (https://orcid.org/0000-0002-5184-9413) [researcher, conceptor, contributor]
  \item Yan Zhu \email{yzhu015@ucr.edu} (https://orcid.org/0000-0002-5952-2108) [researcher, conceptor, contributor]
  \item Hoang Dau \email{hdau001@ucr.edu} (https://orcid.org/0000-0003-2439-5185) [researcher, conceptor, contributor]
}

}
\keyword{internal}
