% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsmp-package.R
\docType{package}
\name{tsmp-package}
\alias{tsmp}
\alias{tsmp-package}
\title{tsmp: Time Series with Matrix Profile}
\description{
This package is a toolkit with several functions implementing the Matrix Profile concept that was created by several researchers at CS-UCR <http://www.cs.ucr.edu/~eamonn/MatrixProfile.html>.
    Some code was written by reading the papers, others were author's matlab code ported to R with author's permittion.
}
\details{
Details in progress
}
\references{
\url{http://www.cs.ucr.edu/~eamonn/MatrixProfile.html}
}
\author{
\strong{Maintainer}: Francisco Bischoff \email{fbischoff@med.up.pt}

}
\keyword{internal}
