% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize}
\alias{visualize}
\title{Plots an object generated from one of the algorithms. In some cases multiple plots will be generated}
\usage{
visualize(obj)
}
\arguments{
\item{obj}{dict_like The object to plotddd.}
}
\value{
A list of matplotlib figures
}
\description{
Plots an object generated from one of the algorithms. In some cases multiple plots will be generated
}
\examples{

}
\seealso{
Other main api: 
\code{\link{analyze}()},
\code{\link{compute}()}
}
\concept{main api}
