% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize}
\alias{visualize}
\title{Plots an object generated from one of the algorithms. In some cases multiple plots will be generated}
\usage{
visualize(profile)
}
\arguments{
\item{profile}{PMP or MP data structure to visualize.}
}
\value{
empty
}
\description{
Plots an object generated from one of the algorithms. In some cases multiple plots will be generated
}
\examples{
result <- compute(mp_toy_data$data[, 1], 80)
visualize(result)
}
\seealso{
Other Main API: 
\code{\link{analyze}()},
\code{\link{compute}()}
}
\concept{Main API}
