% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize}
\alias{visualize}
\title{Plots an object generated from one of the algorithms. In some cases multiple plots will be generated}
\usage{
visualize(profile)
}
\arguments{
\item{profile}{a \code{MatrixProfile} or \code{PMP} object.}
}
\description{
Plots an object generated from one of the algorithms. In some cases multiple plots will be generated
}
\examples{
result <- compute(mp_toy_data$data[, 1], 80)
visualize(result)
}
\references{
Website: \url{http://www.cs.ucr.edu/~eamonn/MatrixProfile.html}
}
\seealso{
Other Main API: 
\code{\link{analyze}()},
\code{\link{compute}()},
\code{\link{discords}()},
\code{\link{motifs}()}
}
\concept{Main API}
