% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{write}
\alias{write}
\alias{write.MatrixProfile}
\alias{write.PMP}
\title{Write a TSMP object to JSON file.}
\usage{
write(x, ...)

\method{write}{MatrixProfile}(x, file, ...)

\method{write}{PMP}(x, file, ...)
}
\arguments{
\item{x}{a \code{MatrixProfile} or \code{PMP} object. If not, the \code{base::write()} function will be called.}

\item{...}{other arguments to be passed forward.}

\item{file}{a \code{character} string with the output filename.}
}
\description{
Write a TSMP object to JSON file.
}
\examples{

result <- compute(mp_toy_data$data[, 1], 80)
\donttest{
write(result, file = file.path(tempdir(), "output.json"))
}
}
